Singletons with global objects
made it so UsersService and HubService are singletons containing the current user and the current hub, respectively.
in order for everything to work properly, the current hub MUST get set. This only currently happens after creating a hub because joining is not yet implemented
when joining gets implemented we need to make sure that we are updating the current hub.
sorting now works better because it takes into consideration the time in which the song was added.
Still need to put a blocker on letting the first elelment of the queue get moved
Users tab is currently blank because we are not yet properly setting the users in the database