made sure videos going bye bye after finish play
first problem: forgot that I removed hub UID's because their names are already unique
second problem: remove song would get removed from database and front end would adapt acordingly, but if it was the last song in the list then the text would not take over the video like we want because of race condition.

Fix: check if there is only one song in the list before removal then set the hasSongs boolean after the removal happens so that the view updates accordingly. Technically could still mess up if someone is adding exactly as the last song is playing, but it would be really hard to produce this result.