Fix up voting restrictions
voting should be a one time poll to the database and therefore we do not want to put an observable/subscription on a block of code that will update the vote values every time the database changes.

Also went ahead and let it add the settings files for firebase hosting.