can now remove users from a hub
This commit also does quite a bit of clean up
We need to make sure we are not subscribing to things that we don't need to constantly listen to
Subscribing to something like creating a DB object just makes that object get regenerated every time we delete it

Hub Creator is able to remove users from the hub.
upon the user being removed from the hub, s/he is taken back to the join/create screen and is prompted with a message letting them know that they have been kicked out of the group.