Working on making callbacks back to the initial function call in html
I am making callbacks starting from the songVoted function I made in
hub-main.component.ts, then isSongUpvoted, and then accessing that callback
in the original isSongUpvoted call in the html.

I am not able to find out a way to store a local boolean variable in the
html to finally decide whether the song was upvoted or not.

I feel I am overcomplicating this, but I am not sure how else to do it.