Touch-ups, notifs, Medal enums, init medal page
forgot to make multiple commits so there is a lot going on in this one

Notifications page is setup to show notifications for friend request, achievements and kicks, but achievements/medals and kicks are not fully implemented yet.
Enum files and values created to differentiate the different types of medals. Hoping that this will allow me to keep all general medal data in memory so as to not user that space in the database.
Empty medals page added and routed
Some minor aesthetic touchups throughout
Removed needless print statements througout.