implementation of medals
Enumerated values for medals so they can be in memory instead of taking up space in the database
Objects for interaction of each type of medal
General medal object for controling types of medals
Scores and names given to each medal
Intervals of enum values given in a way that will allow for more medals to be added at a later time without overlapping values nor needed change of code implementation
Single page keeping tabbed interface for viewing all medals, only owned medals or a category of medals

Yet to implement for medals:
	Owned Medals
	Searching Medals
	Medal Profile Page
	Asigning score of medal to user and/or hub

Large commit caused by all of the default Medal data. Not much was changed to the logic of any of the code, just constantvalues created and interfaces for the enumerations of the medals so as to make sure that all of the medal data is kept in memory and not wasting space in database.