rough one, a lot been goin on
everything regarding medals received a full 360
	instead of a strict number enumeration i took advantage of the javascript-json string enumeration ability so that each medal enumeration can be represented by two characters. this will reduce database size by 1 cahracter every time a medal is stored in the database. not huge, but enough of a difference to implement it.

	medals now have arrays and maps to make sure constant look up is available as well as ngFor capability (ngFor can only iterate over arrays)
		We don't care that it takes more lines of code and in-memory storage because we have no chance of every coming within 50% capacity of firebase hosting storage.

figured out how to use material pop-up windows and implemented that for each medal description instead of havinga fully separate component generated with its html and css like the rest of the components

each medal had its own object and enumeration so that it can be a type and a class with accessible members.

some fixes with friends tab of profile page because i didn't quite have the friend's score mapped to memory correctly

notifications page is not listening for changes to the number of friends that a member has and displaying a notification of a medal being received any time the user meets a milestone number of friends

just a notice that these enumerations of the medals make all coding related to the medals trivial. time-consuming and terribly boring to type out, but it requires no logic past making sure your type in the correct enumerated values. This makes it very easy to read and nice for autocompletion when typing into a decent IDE

there was probably more I missed. I'll try to do better with the committing.